requirements.txt
factordb_pycli==1.3.0
sympy==1.11.1
